staticx: Refactor into package
